fix: update build script

+++ custom-ts/scripts/buildOperations.ts
@@ -69,15 +69,18 @@ async function readDir(rootPath: string, ...dirPath: string[]) {
 }
 
 async function writeTsOperationsConfig() {
-  const operations = await readDir(join(__dirname, '../operations'))
-  const tsOperationConfigDir = join(__dirname, '../../exported/.operations')
-  rmSync(tsOperationConfigDir, { force: true, recursive: true })
-  await mkdir(tsOperationConfigDir)
-  for (const operation of operations) {
-    const dirPath = operation.tsPath.split('/')
-    const fileName = dirPath.pop()!
-    await mkdir(join(tsOperationConfigDir, ...dirPath), { recursive: true })
-    await writeFile(join(tsOperationConfigDir, ...dirPath, `${parse(fileName).name}.config.json`), JSON.stringify(operation, null, 2), 'utf-8')
+  const operationsDirPath = join(__dirname, '../operations')
+  if (existsSync(operationsDirPath)) {
+    const operations = await readDir(operationsDirPath)
+    const tsOperationConfigDir = join(__dirname, '../../exported/.operations')
+    rmSync(tsOperationConfigDir, { force: true, recursive: true })
+    await mkdir(tsOperationConfigDir)
+    for (const operation of operations) {
+      const dirPath = operation.tsPath.split('/')
+      const fileName = dirPath.pop()!
+      await mkdir(join(tsOperationConfigDir, ...dirPath), { recursive: true })
+      await writeFile(join(tsOperationConfigDir, ...dirPath, `${parse(fileName).name}.config.json`), JSON.stringify(operation, null, 2), 'utf-8')
+    }
   }
 }
